requirements.txt
# Core dependencies
typer>=0.9.0
rich>=13.5.0
pydantic>=2.0.0
pyyaml>=6.0
tqdm>=4.66.0

# Document processing
PyPDF2>=3.0.0
python-docx>=0.8.11
markdown>=3.4.0

# Code analysis
pylint>=2.17.0
radon>=6.0.0

# Embedding and vector storage
sentence-transformers>=2.2.2
numpy>=1.24.0
scipy>=1.10.0
faiss-cpu>=1.7.4
chromadb>=0.4.0

# API integration
requests>=2.31.0

# Utilities
pathlib>=1.0.1
loguru>=0.7.0
click>=8.1.0